refactor: util과 Validator 분리
