+++ src/constants/message.js
@@ -15,7 +15,7 @@ const OUTPUT_MESSAGE = {
 const ERROR_MESSAGE = {
   OVER_NAME_LENGTH: `자동차 이름은 ${CAR_RULE.MAX_LENGTH}자 이하만 가능합니다.`,
   EMPTY_NAME: '자동차 이름은 1자 이상 입력해야합니다.',
-  NAME_CHARACTER: '자동차 이름은 알파벳 또는 한글로 입력해야합니다.',
+  NAME_CHARACTER: '자동차 이름은 알파벳이나 한글을 공백없이 쉼표로 구분하여 입력해주세요.',
   NAME_SEPARATOR: '자동차 이름은 쉼표로 구분해야합니다.',
   MOVING_COUNT: '이동 횟수는 자연수를 입력해야합니다.',
   DUPLICATED_CAR_NAME: '중복된 자동차 이름이 있습니다.',
